🚚 move server apps into framework/serverApps
