+++ test/e2e/lib/framework/serverApps/mock.ts
@@ -1,7 +1,7 @@
 import cors from 'cors'
 import express from 'express'
 import * as url from 'url'
-import { buildLogs, buildNpm, buildRum, Endpoints } from './sdkBuilds'
+import { buildLogs, buildNpm, buildRum, Endpoints } from '../sdkBuilds'
 
 export function createMockServerApp(endpoints: Endpoints, setup: string) {
   const app = express()
